Combine the lambda parameter type tests into one file
Merge LambdaBodyNestedInvocation, LambdaParamTypeNestedLambda, and
LambdaParamTypeNotCached into LambdaParamTypeFromInference, each as a
static nested class.  Nesting keeps the differing `map` overloads in
separate scopes, so the calls in each test resolve as they did before.

LambdaParamTypeNotCached duplicated LambdaParamTypeNestedLambda's
helpers and dereference test, so those two became a single nested class
NestedLambda; the block-bodied variant is kept as dereferenceBlockBody.
